requirements.txt
streamlit==1.38.0
pydantic==2.9.2
requests==2.32.3
python-dotenv==1.0.1
pandas==2.2.3
sqlalchemy==2.0.35
langchain==0.2.16
langchain-openai==0.1.25
openai==1.45.0